+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Display Synthetic USD balance in the account overview screen.
+
 ## [1.2.3] - 2023-08-28
 
-- Add synthetic usd feature.
+- Add synthetic USD feature.
 - Fix delayed position update.
 - Change contract duration to 7 days.
 - Add settings in coordinator to make contract fee rate configurable during runtime.